refactor : fix some test names and move some test
Moving `channel_id` test from `lampo_cln_test` to `close_channel.rs`
and rename some test names.

Follow up to https://github.com/vincenzopalazzo/lampo.rs/pull/198
